refactor: convert useSerializer composable to atomic util functions
- composable not needed anymore since its not dependent on codeStore
