+++ frontend/src/pages/AppContainer.vue
@@ -8,7 +8,7 @@ import { useRoute } from "vue-router"
 import { usePageMeta } from "frappe-ui"
 
 import { findPageWithRoute } from "@/utils/helpers"
-import { useSerializer } from "@/utils/useSerializer"
+import { jsonToJs, getBlockInstance } from "@/utils/serializer"
 import AppComponent from "@/components/AppComponent.vue"
 
 import useAppStore from "@/stores/appStore"
@@ -23,8 +23,6 @@ const codeStore = useCodeStore()
 const page = ref<StudioPage | null>(null)
 
 const rootBlock = ref<Block | null>(null)
-
-const { jsonToJs, getBlockInstance } = useSerializer()
 
 watch(
 	() => route.path,
